requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
coverage==6.5.0
cryptography==43.0.3
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.51.0
djangorestframework==3.15.0
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.0
pytest==7.3.1
pytest-django==4.5.2
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
schema==0.7.7
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
